+++ src/jsx-dev-runtime.ts
@@ -7,16 +7,21 @@ import { createJSXVNode } from "./normalize.ts";
 import type { ComponentType, JsxDevSource, Key, VNode } from "./types.ts";
 import { Fragment } from "./types.ts";
 
-import type { JsxElement, JsxKey, JsxProps, JsxRenderable } from "./jsx-types.ts";
+import type { JsxKey, JsxProps, JsxRenderable } from "./jsx-types.ts";
 
 export { Fragment };
 export type { JsxDevSource };
-
-export type { JsxElement, JsxKey, JsxProps, JsxRenderable } from "./jsx-types.ts";
+export type {
+  JsxElement,
+  JsxKey,
+  JsxProps,
+  JsxRenderable,
+  JsxTag,
+} from "./jsx-types.ts";
 
 // deno-lint-ignore no-namespace
 export namespace JSX {
-  export interface Element extends JsxElement {}
+  export type Element = VNode;
 
   export interface ElementChildrenAttribute {
     children: unknown;
@@ -29,6 +34,10 @@ export namespace JSX {
   export interface IntrinsicElements {
     [name: string]: JsxProps;
   }
+
+  export type ElementType =
+    | keyof IntrinsicElements
+    | ((props: Record<string, unknown>) => Element);
 
   export type Child = JsxRenderable;
 }